# Improve DFS performance (fix #1312)