I ran this benchmark on `g = star_digraph(100_000)`.

Old version:
```
julia> @benchmark LightGraphs.Experimental.Traversals.is_cyclic(g)
BenchmarkTools.Trial: 
  memory estimate:  110.42 KiB
  allocs estimate:  8
  --------------
  minimum time:     7.407 s (0.00% GC)
  median time:      7.407 s (0.00% GC)
  mean time:        7.407 s (0.00% GC)
  maximum time:     7.407 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1
```

New version:
```
julia> @benchmark LightGraphs.Experimental.Traversals.is_cyclic(g)
BenchmarkTools.Trial: 
  memory estimate:  2.87 MiB
  allocs estimate:  25
  --------------
  minimum time:     1.814 ms (0.00% GC)
  median time:      1.878 ms (0.00% GC)
  mean time:        1.916 ms (0.59% GC)
  maximum time:     2.678 ms (12.22% GC)
  --------------
  samples:          2591
  evals/sample:     1
```