@estelle0500 2020-03-02
I could use some input on where the visitor functions should be called inside the DFS traversal. I moved `newvisitfn!` so that it would behave like recursive DFS, and I am trying to figure out if the other visitor functions are correctly placed. Also, I have no idea where to put `postlevelfn!` - the concept of a level just doesn't seem to fit in with DFS.