@estelle0500 2020-03-02
I also want to add documentation that says that the DFS function explores the neighbors of each vertex in _reverse order_. This probably isn't very intuitive for users. In fact, the tests expected the vertices to be explored in the usual forward order, which is why I had to modify the tests. Where's a good place to document this behavior?