@sbromberger 2020-03-02
This doesn't seem to produce the correct result on a `wheel_graph(10)`.