@estelle0500 2020-03-03
Yes, I see that it isn't working for `wheel_graph(10)`. I'll try to fix this and add a test suite for `is_cyclic` (paths, cycles, etc).