@sbromberger 2020-03-03
@estelle0500 - let's take a look at #1321 - this is a version that's closer to the original code and appears to run correctly (and marginally faster). Can we adapt it so that it can be used for `Johnson`?

I don't want to discount all the work you've done in this PR, but this is turning out to be a pretty radical change to the existing algorithm and I'd like to be conservative where possible.

Thoughts?