@estelle0500 2020-03-04
Yes, I'm pretty sure #1321 is correct - this is the same thing I came up with while I was thinking of why my implementation of DFS failed on `wheel_graph` and how to fix it. We can go ahead and merge #1321 after it's cleaned up. In that case, I'll rebase my commits (for adding multiple dispatch visitor functions) and combine those with later commits for Johnson's cycle-finding code in one PR.